requirements.txt
unstructured
tiktoken
pinecone
pypdf
openai
langchain
pandas
numpy
python-dotenv
accelerate